fix(core): 修复类型转换以确保安全返回整数值
- 在Array.hpp中对size()和index()的返回值进行了int类型转换
- 将unshift()和push()方法的返回值转换为int32_t类型
- 在RSASigner.hpp中新建内存缓冲区时转换keyLen为int类型以防止类型不匹配
- 通过静态类型转换确保数据类型安全和代码健壮性

+++ include/darabonba/Array.hpp
@@ -31,10 +31,10 @@ public:
   static int index(const std::vector<std::string> &raw,
                    const std::string &str) {
     auto pos = std::find(raw.begin(), raw.end(), str);
-    return pos == raw.end() ? -1 : pos - raw.begin();
+    return pos == raw.end() ? -1 : static_cast<int>(pos - raw.begin());
   }
 
-  static int size(const std::vector<std::string> &raw) { return raw.size(); }
+  static int size(const std::vector<std::string> &raw) { return static_cast<int>(raw.size()); }
 
   static const std::string &get(const std::vector<std::string> &raw,
                                 int index) {
@@ -112,7 +112,7 @@ public:
   static int32_t unshift(std::vector<std::string> &arr,
                          const std::string &element) {
     arr.insert(arr.begin(), element);
-    return arr.size();
+    return static_cast<int32_t>(arr.size());
   }
 
   // Push: Adds one or more elements to the end of an array and returns the new
@@ -120,7 +120,7 @@ public:
   static int32_t push(std::vector<std::string> &arr,
                       const std::string &element) {
     arr.push_back(element);
-    return arr.size();
+    return static_cast<int32_t>(arr.size());
   }
 
   // Ascending Sort: Sorts the elements of an array in ascending order
